requirements.txt
django
django-crispy-forms
bootstrap4
requests
yfinance
lxml
PyGithub
pocketcasts-api